Changes to be committed: 	modified:   package-lock.json 	modified:   package.json 	new file:   precision_classify.csv 	modified:   src/app.d.ts 	modified:   src/lib/components/Footer.svelte 	modified:   src/lib/components/Navigation.svelte 	new file:   src/lib/modal/EmailVerification.svelte 	new file:   src/lib/modal/TermsAndCondition.svelte 	modified:   src/lib/modal/ViewPicture.svelte 	modified:   src/routes/+layout.svelte 	modified:   src/routes/+page.svelte 	new file:   src/routes/api/user/send-verification/+server.ts 	modified:   src/routes/folder/[slug]/+page.svelte 	modified:   src/routes/live/+page.svelte 	modified:   src/routes/login/+page.svelte 	modified:   src/routes/login/loginform.svelte 	modified:   src/routes/record/+page.svelte 	modified:   src/routes/spray/spraybuttons.svelte 	modified:   src/routes/wifi/+page.svelte 	modified:   test2.py 	new file:   test3.py 	new file:   test4.py 	new file:   test5.py

+++ src/lib/modal/ViewPicture.svelte
@@ -7,6 +7,7 @@
 	export let downloadImage;
 	export let selectedImage;
 	export let images: any;
+	export let noDownloadDelete = false;
 
 	async function deleteImage(public_id: string) {
 		const toastId = addToast('Deleting image...', 'loading');
@@ -36,11 +37,11 @@
 
 {#if modalOpen}
 	<div class="fixed inset-0 z-50 flex items-center justify-center bg-black/80 p-4">
-		<!-- conditionally change width -->
 		<div
-			class={`relative w-full ${selectedImage.plantName === 'SCANBOX' ? 'max-w-3xl' : 'max-w-lg'} 
-                    rounded-lg border border-gray-300 bg-white shadow-xl 
-                    dark:border-gray-700 dark:bg-gray-900`}
+			class={`relative w-full 
+                ${selectedImage.plantName === 'SCANBOX' ? 'max-w-5xl' : 'max-w-6xl'} 
+                flex rounded-xl border border-gray-300 bg-white 
+                shadow-xl dark:border-gray-700 dark:bg-gray-900`}
 		>
 			<button
 				class="absolute top-3 right-3 rounded-full bg-gray-200 p-2 text-gray-700 hover:bg-gray-300 dark:bg-gray-700 dark:text-white dark:hover:bg-gray-600"
@@ -49,21 +50,27 @@
 				✖
 			</button>
 
-			<div class="relative">
+			<div
+				class="relative flex w-1/2 items-center justify-center rounded-xl bg-gray-300 p-2 dark:bg-gray-800"
+			>
 				<img
-					class={`w-full rounded-t-lg object-contain 
-                           ${selectedImage.plantName === 'SCANBOX' ? 'max-h-[500px]' : 'max-h-[500px]'}`}
+					class="max-h-[600px] w-full rounded-xl object-contain"
 					src={selectedImage.src}
 					alt="Zoomed"
 				/>
 				<div
-					class="absolute bottom-0 left-0 w-full bg-gray-300/50 p-2 text-center text-sm text-white dark:bg-gray-900/50"
+					class="absolute bottom-0 left-0 w-full bg-gray-900/50 p-4 text-center text-sm text-white dark:bg-gray-900/50"
 				>
 					{selectedImage.id} | Captured at: {selectedImage.timestamp}
 				</div>
 			</div>
 
-			<div class="space-y-2 p-4">
+			<div class="max-h-[600px] w-1/2 space-y-3 overflow-y-auto p-6">
+				<h2
+					class="mb-4 border-b border-green-500/40 pb-2 text-xl font-semibold text-gray-800 dark:text-green-200"
+				>
+					{selectedImage.plantName === 'SCANBOX' ? 'SCANBOX' : 'ROI PLANT'}
+				</h2>
 				{#if selectedImage.plantName !== 'SCANBOX'}
 					<p class="text-sm text-gray-600 dark:text-gray-400">
 						<strong>Plant:</strong>
@@ -78,9 +85,9 @@
 						{selectedImage.imageSize}
 					</p>
 					<p
-						class="max-h-40 overflow-y-auto text-sm whitespace-pre-wrap text-gray-600 dark:text-gray-400"
+						class="max-h-100 overflow-y-auto text-sm whitespace-pre-wrap text-gray-600 dark:text-gray-400"
 					>
-						<strong>AI Analysis:</strong><br/>{selectedImage.generatedDescription}
+						<strong>AI Analysis:</strong><br />{selectedImage.generatedDescription}
 					</p>
 				{:else}
 					<p class="text-sm text-gray-600 dark:text-gray-400">
@@ -89,26 +96,30 @@
 					</p>
 				{/if}
 
-				<button
-					class="mt-2 w-full rounded-lg bg-green-600 px-4 py-2 text-white hover:bg-green-700 dark:bg-green-700 dark:hover:bg-green-800"
-					on:click={() => downloadImage(selectedImage.src)}
-				>
-					Download Image
-				</button>
+				<div class="space-y-2">
+					{#if !noDownloadDelete}
+						<button
+							class="w-full rounded-lg bg-green-600 px-4 py-2 text-white hover:bg-green-700 dark:bg-green-700 dark:hover:bg-green-800"
+							on:click={() => downloadImage(selectedImage.src)}
+						>
+							Download Image
+						</button>
 
-				<button
-					class="w-full rounded-lg bg-red-500 px-4 py-2 text-white hover:bg-red-600 dark:bg-red-800 dark:hover:bg-red-900"
-					on:click={() => deleteImage(selectedImage.id)}
-				>
-					Delete Image
-				</button>
+						<button
+							class="w-full rounded-lg bg-red-500 px-4 py-2 text-white hover:bg-red-600 dark:bg-red-800 dark:hover:bg-red-900"
+							on:click={() => deleteImage(selectedImage.id)}
+						>
+							Delete Image
+						</button>
+					{/if}
 
-				<button
-					class="w-full rounded-lg bg-gray-400 px-4 py-2 text-gray-900 hover:bg-gray-500 dark:bg-gray-700 dark:text-white dark:hover:bg-gray-800"
-					on:click={closeModal}
-				>
-					Close
-				</button>
+					<button
+						class="w-full rounded-lg bg-gray-400 px-4 py-2 text-gray-900 hover:bg-gray-500 dark:bg-gray-700 dark:text-white dark:hover:bg-gray-800"
+						on:click={closeModal}
+					>
+						Close
+					</button>
+				</div>
 			</div>
 		</div>
 	</div>
